+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - System setting for performer liability insurance question
 - System setting to configure acceptable back-to-back intervals
 - Last update information on performer scheduling selections
+- Require day of event contact information when scheduling performer
 - End date to streaming program list on dashboard and streaming page
 - Scheduled task execution capability
 - Welcome email capability